+++ R/hooks.R
@@ -31,8 +31,6 @@ print.rglId <- function(x, rglwidget = getOption("rgl.printRglwidget", FALSE),
     # FIXME:  For lowlevel, this should replace the scene, not update the history
     print(rglwidget(...))
   else if (in_pkgdown_example())
-    pkgdown_print(x)
-  else if (in_pkgdown()) # Must not have pkgdown_print defined
-    cat("")
+    pkgdown::pkgdown_print(x)
   invisible(x)
 }